End of video 11 commit

+++ app/Http/Controllers/PostsController.php
@@ -6,8 +6,12 @@ use Illuminate\Http\Request;
 use App\Post;
 use Illuminate\Support\Facades\DB;
 
-class PostsController extends Controller
-{
+class PostsController extends Controller {
+
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -75,6 +79,12 @@ class PostsController extends Controller
      */
     public function edit($id) {
         $post = Post::find($id);
+
+        //Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -108,6 +118,12 @@ class PostsController extends Controller
      */
     public function destroy($id) {
         $post = Post::find($id);
+
+        //Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed');
     }
